refactor: moved the insert query for booking from route handler to bookingDao
