+++ lib/daos/bookingDao.js
@@ -1,4 +1,5 @@
 import { bookings } from 'models';
+import { sequelize } from 'models';
 
 const attributes = ['id', 'fromLocation', 'toLocation', 'byUser'];
 
@@ -9,13 +10,34 @@ const attributes = ['id', 'fromLocation', 'toLocation', 'byUser'];
  */
 export const createOneBooking = async data => {
     const { toLocation, fromLocation, userId } = data;
-    const booking = await bookings.create({
-        toLocation,
-        fromLocation,
-        byUser: userId
-    });
 
-    return booking;
+    // The Lat and Lng will be sent as a stringified array
+    // for both to and from location, which can then be parsed
+    // and be used to reference and create points
+    const parsedToLocation = JSON.parse(toLocation);
+    const parsedFromLocation = JSON.parse(fromLocation);
+
+    await sequelize.query(
+        `
+        INSERT INTO bookings VALUES
+            (
+                DEFAULT,
+                ST_SRID(ST_GeomFromText('POINT(:fromLat :fromLng)'), 4326),
+                ST_SRID(ST_GeomFromText('POINT(:toLat :toLng)'), 4326),
+                :userId,
+                DEFAULT
+            )
+        `,
+        {
+            replacements: {
+                fromLat: parsedFromLocation[0],
+                fromLng: parsedFromLocation[1],
+                toLat: parsedToLocation[0],
+                toLng: parsedToLocation[1],
+                userId
+            }
+        }
+    );
 };
 
 /**
