+++ lib/routes/booking/routes.js
@@ -1,7 +1,6 @@
 import { badImplementation } from '@hapi/boom';
-import { sequelize } from 'models';
 import Joi from '@hapi/joi';
-import { findOneBooking } from 'daos/bookingDao';
+import { createOneBooking, findOneBooking } from 'daos/bookingDao';
 import { numberSchema, stringSchema } from 'utils/validationUtils';
 
 module.exports = [
@@ -47,33 +46,7 @@ module.exports = [
         handler: async (request, h) => {
             try {
                 const { toLocation, fromLocation, userId } = request.payload;
-
-                // The Lat and Lng will be sent as a stringified array
-                // for both to and from location, which can then be parsed
-                // and be used to reference and create points
-                const parsedToLocation = JSON.parse(toLocation);
-                const parsedFromLocation = JSON.parse(fromLocation);
-
-                await sequelize.query(
-                    `
-                    INSERT INTO bookings VALUES
-                     (
-                        DEFAULT,
-                        ST_SRID(ST_GeomFromText('POINT(:fromLat :fromLng)'), 4326),
-                        ST_SRID(ST_GeomFromText('POINT(:toLat :toLng)'), 4326),
-                        :userId
-                     )
-                `,
-                    {
-                        replacements: {
-                            fromLat: parsedFromLocation[0],
-                            fromLng: parsedFromLocation[1],
-                            toLat: parsedToLocation[0],
-                            toLng: parsedToLocation[1],
-                            userId
-                        }
-                    }
-                );
+                await createOneBooking({ toLocation, fromLocation, userId });
 
                 return h.response({ status: 'Booking created successfully!' });
             } catch (error) {
