+++ resources/v2/01_bookings.sql
@@ -4,5 +4,6 @@ CREATE TABLE bookings
     from_location   POINT NOT NULL SRID 4326,
     to_location     POINT NOT NULL SRID 4326,
     by_user         INT NOT NULL,
+    created_at      DATETIME DEFAULT CURRENT_TIMESTAMP NOT NULL, 
     CONSTRAINT bookings_users_fk FOREIGN KEY (by_user) REFERENCES users (id) ON UPDATE CASCADE
   )
